==> 1. Making_Database.ipynb <==
import os
import sqlite3
import pandas as pd

# 1. DB 만들기

db_name = 'Kickstarter.db'

# db 생성
con = sqlite3.connect( db_name )
cur = con.cursor()

# 2. Table 생성
- 2.1 Creator
- 2.2 Update
- 2.3 Project
- 2.4 Img
- 2.5 Thumbnail
- 2.6 Video

## 2.1 Creator

# ON DELETE CASCADE : if the parent record is deleted, any child records are also deleted
# ON UPDATE CASCADE : if the parent primary key is change, the child value will also change to refelct that
# ON UPDATE CASCADE ON DELETE CASCADE : if you UPDATE OR DELETE the parent, the change is cascaded to the child

sql =  ''' Create Table if not exists Creator(
  Creator_Id  varchar(20)  NOT NULL,
  Creator_Name varchar(20)  NULL,
  Signed_up_date varchar(50)  NULL,
  Backed_count TEXT  NULL,
  Created_count INT(10)  NULL,
  Comment_count varchar(10)  NULL,
  Profile_url varchar(255)  NULL,
  Creator_url varchar(255)  NULL,
  PRIMARY KEY (Creator_Id)
        ) '''
cur.execute(sql)

## 2.2 Update_project

sql =  ''' Create Table if not exists Upd(
  Project_Id varchar(20)  NOT NULL,
  Creator_Id varchar(20)  NOT NULL,
  Date varchar(50)  NULL,
  Days_to_go varchar  NULL,
  Pledged varchar(20)  NULL,
  Backers varchar(10)  NULL,
  Options_number INT(10)  NULL,
  Update_staff_pic INT(10)  NULL,
  Title_count INT(20)  NULL,
  Blurb_count INT(20)  NULL,
  Content_text_length INT(20)  NULL,
  Content_img_count INT(20)  NULL,
  Faq_count INT(10)  NULL,
  Update_count INT(10)  NULL,
  Comment_count INT(10)  NULL,
  FOREIGN KEY (Creator_Id)
      REFERENCES Creator (Creator_Id) ON DELETE CASCADE  ON UPDATE CASCADE
  FOREIGN KEY (Project_Id)
      REFERENCES Project (Project_Id) ON DELETE CASCADE  ON UPDATE CASCADE    
        ) ''' 
cur.execute(sql)

## 2.3 Project

# ON DELETE CASCADE : if the parent record is deleted, any child records are also deleted
# ON UPDATE CASCADE : if the parent primary key is change, the child value will also change to refelct that
# ON UPDATE CASCADE ON DELETE CASCADE : if you UPDATE OR DELETE the parent, the change is cascaded to the child

sql =  ''' Create Table if not exists Project(
  Project_Id  varchar(20)  NOT NULL,
  Date FLOAT(20)  NULL,
  Project_title TEXT(50)  NULL,
  Project_blurb TEXT  NULL,
  Project_goal FLOAT(10)  NULL,
  Project_deadline INT(10)  NULL,
  Project_launched_at INT(20)  NULL,
  Project_country varchar(20)  NULL,
  Project_currency varchar(10)  NULL,
  Project_currency_symbol varchar(10)  NULL,
  Project_category varchar(20)  NULL,
  Project_local varchar(20)  NULL,
  Cancel_check INT(20)  NULL,
  PRIMARY KEY (Project_Id)
        ) '''
cur.execute(sql)

## 2.4 Video

# ON DELETE CASCADE : if the parent record is deleted, any child records are also deleted
# ON UPDATE CASCADE : if the parent primary key is change, the child value will also change to refelct that
# ON UPDATE CASCADE ON DELETE CASCADE : if you UPDATE OR DELETE the parent, the change is cascaded to the child

sql =  ''' Create Table if not exists Video(
  Project_Id  varchar(20)  NOT NULL,
  Video_name varchar(255)  NULL,
  FOREIGN KEY (Project_Id)
      REFERENCES Project (Project_Id) ON DELETE CASCADE  ON UPDATE CASCADE    
        ) '''
cur.execute(sql)

## 2.5 Image

# ON DELETE CASCADE : if the parent record is deleted, any child records are also deleted
# ON UPDATE CASCADE : if the parent primary key is change, the child value will also change to refelct that
# ON UPDATE CASCADE ON DELETE CASCADE : if you UPDATE OR DELETE the parent, the change is cascaded to the child

sql =  ''' Create Table if not exists Img(
  Creator_Id  varchar(20)  NOT NULL,
  Creator_Name varchar(20)  NULL,
  Signed_up_date varchar(50)  NULL,
  Backed_count TEXT  NULL,
  Created_count INT(10)  NULL,
  Comment_count INT(10)  NULL,
  Profile_url varchar(255)  NULL,
  Creator_url INT(10)  NULL,
  PRIMARY KEY (Creator_Id)
        ) '''
cur.execute(sql)

## 2.6 Thumbnail

# ON DELETE CASCADE : if the parent record is deleted, any child records are also deleted
# ON UPDATE CASCADE : if the parent primary key is change, the child value will also change to refelct that
# ON UPDATE CASCADE ON DELETE CASCADE : if you UPDATE OR DELETE the parent, the change is cascaded to the child

sql =  ''' Create Table if not exists Thumbnail(
  Creator_Id  varchar(20)  NOT NULL,
  Creator_Name varchar(20)  NULL,
  Signed_up_date varchar(50)  NULL,
  Backed_count TEXT  NULL,
  Created_count INT(10)  NULL,
  Comment_count INT(10)  NULL,
  Profile_url varchar(255)  NULL,
  Creator_url INT(10)  NULL,
  PRIMARY KEY (Creator_Id)
        ) '''
cur.execute(sql)
