==> 2. Insert_data.ipynb <==
import pandas as pd
import os
import sqlite3

db_FILE = 'Kickstarter.db'
con = sqlite3.connect( db_FILE )
cur = con.cursor()

## Creator

# Creator
creator_csv = './data/creator_table.csv'
creator_df = pd.read_csv(creator_csv,encoding='utf-16')
creator_df = creator_df.drop('profile_img',axis=1)
unique_creator_df = creator_df.drop_duplicates(subset=['creator_id'],keep='last')
unique_creator_df.to_sql('Creator',con,if_exists='append',index=False)

# cursor = con.execute('select * from Upd')
# names = list(map(lambda x: x[0], cursor.description))
# names = [description[0] for description in cursor.description]
# names

## Update

# Upd
update_csv = 'new_update_df.csv'
update_df = pd.read_csv(update_csv,encoding='utf-16')

update_df.to_sql('Upd',con,if_exists='append',index=False)

## Project

# Project
project_csv = './data/project_table.csv'
project_df = pd.read_csv(project_csv,encoding='utf-16')
drop_columns = ['creator_id','creator_profile_img','creator_name','creator_url','project_main_img','project_url']
column_names = project_df.columns.tolist()
#new_colimn_names = list(filter(lambda x : x not in removed_item,column_names))
new_project_df = project_df.drop(drop_columns, axis=1) 

new_project_df.to_sql('Project',con,if_exists='append',index=False)

## Video

# Video
video_csv = './data/refined_video_table.csv'
video_df = pd.read_csv(video_csv,encoding='utf-16')

video_df.to_sql('Video',con,if_exists='append',index=False)
