==> 4-3. Video filtering.ipynb <==
import pandas as pd

total_df = total_df[total_df['check'] != '0']

total_df.to_csv('english_video.csv',encoding='utf-16',index=False)

df = pd.read_csv('0803_Final_Data.csv',encoding='utf-16')

en_proj = total_df['project_id'].unique().tolist()

filter_df = df[df['project_id'].isin(en_proj)]

success_df = filter_df[filter_df['success'] == 'success']
fail_df = filter_df[filter_df['success'] == 'Fail']

success_df.shape

fail_df.shape

success_df

success_df_500 = success_df.sample(n=500, random_state=22)
fail_df_500 = fail_df.sample(n=500, random_state=22)

Final_df = pd.concat([success_df_500,fail_df_500],axis=0)

Final_df.to_csv('0803_Final_filtered_df.csv',encoding='utf-16',index=False)
