==> 6. Prediction model.ipynb <==
import pandas as pd
import numpy as np
from Utilities import Model_evaluation

df = pd.read_csv('./0803Data/analysis_data.csv',encoding='utf-16')
df_duration = pd.read_csv('./0803Data/duration_survey_df.csv',encoding='utf-16')

video_df = df_duration[['project_id','duration']]
video_df = video_df.rename(columns = {'duration':'video_duration'})

final_df = pd.merge(df,video_df,how='left')

# 최종 1000개 비디오만 추출
final_video_df = pd.read_csv('./0803Data/0803_Final_filtered_df.csv',encoding='utf-16')
final_list = final_video_df['project_id'].tolist()
final_df = final_df[final_df.project_id.isin(final_list)]

final_df

final_df.to_csv('0803_Final_video_data.csv',index=False,encoding='utf-16')

final_df.shape

final_df.columns.tolist()

Analysis_df = final_df[['duration','project_goal','video_duration','update_count','project_blurb','faq_count',\
    'created_count','backed_count','ratio1','ratio2','ratio3','backer1','backer2','backer3','success']]

# Project_blurb
Analysis_df['project_blurb'] = Analysis_df['project_blurb'].apply(lambda x : len(x))
# Backed_count
Analysis_df['backed_count'] = Analysis_df['backed_count'].apply(lambda x : 0 if x =='None' else x)
Analysis_df['backed_count'] = Analysis_df['backed_count'].apply(lambda x : float(x))

## Model evaluation

model = Model_evaluation.Model_evaluation(Analysis_df,'success',scaler='std')

### Random forest

model.random_forest()

### SVM

model.svm()

### Logistic regression

model.logistic_regression()

### MLP

model.mlp()
